updated footer
