--- src/components/Footer/styles.css
@@ -1,103 +1,0 @@
-.footer {
-    background-color: #E9F3E7;
-    padding: 0px 0;
-    text-align: center;
-    font-family: 'Averia Sans Libre', sans-serif;
-}
-  
-.footer-bottom {
-    background-color: #728F6E;
-    padding: 10px 0;
-    color: white;
-}
-
-.footer-bottom p {
-    margin: 0;
-}
-  
-.nav {
-    display: flex;
-    justify-content: center;
-    gap: 20px;
-    margin-bottom: 10px;
-    font-family: 'Averia Sans Libre', sans-serif;
-    font-size: 12px;
-    color: #3A6351;
-}
-  
-.nav a {
-    text-decoration: none;
-    color: #3A6351;
-    font-weight: bold;
-    transition: color 0.2s ease-in-out;
-}
-  
-.nav a:hover {
-    color: #1A3C34;
-}
-
-.nav-link button {
-    font-size: 12px;
-}
-  
-.copyright {
-    font-family: 'Karma', serif;
-    font-size: 14px;
-    color: white;
-}
-
-.media-links {
-    display: flex; 
-    justify-content: center; 
-    align-items: center;
-    gap: 60px; 
-    padding-top: 30px;
-    padding-bottom: 10px;
-}
-  
-.media-links a {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 54px;
-    height: 54px;
-    background-color: #F56038;
-    border-radius: 50%;
-    color: white;
-    font-size: 24px; 
-    transition: transform 0.2s ease, background-color 0.2s ease;
-}
-  
-.media-links a:hover {
-    transform: scale(1.1);
-    background-color: #D94A25;
-}
-
-
-@media (max-width: 480px) {
-    .nav {
-        gap: 0px;
-    }
-
-    .nav-link {
-        padding: 5px;
-    }
-
-    .nav-link button {
-        font-size: 8px;
-    }
-
-    .media-links a {
-        width: 35px;
-        height: 35px;
-        font-size: 20px; 
-    }
-
-    .media-links {
-        gap: 10px; 
-    }
-
-    .copyright {
-        font-size: 8px;
-    }
-}
